Refactor util for mappings
